Renames OID::Spatial#binary to #binary_string?
Renaming the method solves the issue of the method being called by
ActiveRecord's LogSubscriber, which expects `OID::Spatial` to inherit or
overwrite the `binary?` method of ActiveRecord's `Type::Value`.

If ActiveRecord called the method before this change a NoMethodError was
thrown, since the `binary?` method was private.

This change renames the method since it's for internal use only.